+++ lib/features/release/application/providers/form_release/form_release_notifier.dart
@@ -52,6 +52,12 @@ class ReleaseFormStateNotifier extends StateNotifier<ReleaseFormState> {
     state = ReleaseFormState();
   }
 
+  void updateVersionCode({required String value}) {
+    state = state.copyWith(
+      versionCode: value
+    );
+  }
+
 }
 
 final releaseFormProvider = StateNotifierProvider<ReleaseFormStateNotifier, ReleaseFormState>(