+++ lib/features/release/data/dtos/get_all_releases/release_code_version_dto.dart
@@ -1,0 +1,18 @@
+class ReleaseVersionCodeDto {
+
+  final String id;
+  final String code;
+
+  ReleaseVersionCodeDto({
+    required this.id,
+    required this.code,
+  });
+
+  factory ReleaseVersionCodeDto.fromJson(Map<String, dynamic> json) {
+    return ReleaseVersionCodeDto(
+      id: json["id"] ?? "",
+      code: json["code"] ?? "",
+    );
+  }
+
+}