+++ lib/features/version_codes/presentation/providers/version_codes_provider.g.dart
@@ -1,0 +1,25 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+part of 'version_codes_provider.dart';
+
+// **************************************************************************
+// RiverpodGenerator
+// **************************************************************************
+
+String _$versionCodesHash() => r'4ffa29ebc17f8b2b7da47e25a9a6222b3792b4ee';
+
+/// See also [VersionCodes].
+@ProviderFor(VersionCodes)
+final versionCodesProvider =
+    NotifierProvider<VersionCodes, VersionCodesState>.internal(
+  VersionCodes.new,
+  name: r'versionCodesProvider',
+  debugGetCreateSourceHash:
+      const bool.fromEnvironment('dart.vm.product') ? null : _$versionCodesHash,
+  dependencies: null,
+  allTransitiveDependencies: null,
+);
+
+typedef _$VersionCodes = Notifier<VersionCodesState>;
+// ignore_for_file: type=lint
+// ignore_for_file: subtype_of_sealed_class, invalid_use_of_internal_member, invalid_use_of_visible_for_testing_member, deprecated_member_use_from_same_package